create .gitlab-ci.yml
Also add a status message, announcing when pre-existing results were
found in the working directory (cache). When this happens, an element is
skipped (not checked out). Adding the status message makes it easier to
follow what the script is doing.
Also add the Dockerfiles that were used to build the CI docker images
(for infomation).